# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CompanyResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "CompanyResponse",
        "type": "object"
      },
...
      "DecisionResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "decision_type": {
            "title": "Decision Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "opportunity_id": {
            "title": "Opportunity Id",
            "type": "string"
          },
          "previous_status": {
            "enum": [
              "new",
              "to_review",
              "interesting",
              "shortlisted",
              "deferred",
              "excluded",
              "archived"
            ],
            "title": "Previous Status",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "resulting_status": {
            "enum": [
              "new",
              "to_review",
              "interesting",
              "shortlisted",
              "deferred",
              "excluded",
              "archived"
            ],
            "title": "Resulting Status",
            "type": "string"
          },
          "reverses_decision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reverses Decision Id"
          }
        },
        "required": [
          "id",
          "opportunity_id",
          "decision_type",
          "previous_status",
          "resulting_status",
          "reason",
          "created_at",
          "reverses_decision_id"
        ],
        "title": "DecisionResponse",
        "type": "object"
      },
...
      "ExclusionPayload": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ExclusionPayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpportunityDetailResponse": {
        "properties": {
          "assessments": {
            "items": {
              "$ref": "#/components/schemas/AssessmentResponse"
            },
            "title": "Assessments",
            "type": "array"
          },
          "availability": {
            "anyOf": [
              {
                "enum": [
                  "available",
                  "unavailable",
                  "uncertain",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability"
          },
          "availability_age_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability Age Days"
          },
          "availability_last_checked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability Last Checked At"
          },
          "company": {
            "$ref": "#/components/schemas/CompanyResponse"
          },
          "decision_history": {
            "items": {
              "$ref": "#/components/schemas/DecisionResponse"
            },
            "title": "Decision History",
            "type": "array"
          },
          "external_assessments": {
            "items": {
              "$ref": "#/components/schemas/AssessmentResponse"
            },
            "title": "External Assessments",
            "type": "array"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/OpportunityGroupSummaryResponse"
            },
            "title": "Groups",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "import_provenance": {
            "$ref": "#/components/schemas/ImportProvenanceResponse"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/LocationResponse"
            },
            "title": "Locations",
            "type": "array"
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/ObservationResponse"
            },
            "title": "Observations",
            "type": "array"
          },
          "personal_assessment_history": {
            "items": {
              "$ref": "#/components/schemas/PersonalAssessmentResponse"
            },
            "title": "Personal Assessment History",
            "type": "array"
          },
          "personal_assessments": {
            "items": {
              "$ref": "#/components/schemas/PersonalAssessmentResponse"
            },
            "title": "Personal Assessments",
            "type": "array"
          },
          "postings": {
            "items": {
              "$ref": "#/components/schemas/PostingResponse"
            },
            "title": "Postings",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceResponse"
            },
            "title": "Sources",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tracking_status": {
            "enum": [
              "new",
              "to_review",
              "interesting",
              "shortlisted",
              "deferred",
              "excluded",
              "archived"
            ],
            "title": "Tracking Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "company",
          "locations",
          "postings",
          "sources",
          "observations",
          "tracking_status",
          "external_assessments",
          "assessments",
          "personal_assessments",
          "personal_assessment_history",
          "decision_history",
          "import_provenance"
        ],
        "title": "OpportunityDetailResponse",
        "type": "object"
      },
...
      "OpportunityGroupSummaryResponse": {
        "properties": {
          "group_id": {
            "title": "Group Id",
            "type": "string"
          },
          "group_type": {
            "enum": [
              "general",
              "application_wave"
            ],
            "title": "Group Type",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "group_id",
          "name",
          "group_type"
        ],
        "title": "OpportunityGroupSummaryResponse",
        "type": "object"
      },
      "OpportunityListItemResponse": {
        "properties": {
          "assessment_count": {
            "title": "Assessment Count",
            "type": "integer"
          },
          "availability": {
            "anyOf": [
              {
                "enum": [
                  "available",
                  "unavailable",
                  "uncertain",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability"
          },
          "availability_age_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability Age Days"
          },
          "availability_last_checked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability Last Checked At"
          },
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/OpportunityGroupSummaryResponse"
            },
            "title": "Groups",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "import_id": {
            "title": "Import Id",
            "type": "string"
          },
          "imported_at": {
            "title": "Imported At",
            "type": "string"
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "title": "Locations",
            "type": "array"
          },
          "posting_count": {
            "title": "Posting Count",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tracking_status": {
            "enum": [
              "new",
              "to_review",
              "interesting",
              "shortlisted",
              "deferred",
              "excluded",
              "archived"
            ],
            "title": "Tracking Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "company_name",
          "locations",
          "posting_count",
          "assessment_count",
          "import_id",
          "imported_at",
          "tracking_status"
        ],
        "title": "OpportunityListItemResponse",
        "type": "object"
      },
...
      "PersonalAssessmentPayload": {
        "additionalProperties": false,
        "properties": {
          "criterion_id": {
            "minLength": 1,
            "title": "Criterion Id",
            "type": "string"
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "criterion_id",
          "value"
        ],
        "title": "PersonalAssessmentPayload",
        "type": "object"
      },
      "PersonalAssessmentResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "criterion_id": {
            "title": "Criterion Id",
            "type": "string"
          },
          "criterion_name": {
            "title": "Criterion Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "opportunity_id": {
            "title": "Opportunity Id",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          },
          "revision_number": {
            "title": "Revision Number",
            "type": "integer"
          },
          "supersedes_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Id"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "id",
          "opportunity_id",
          "criterion_id",
          "criterion_name",
          "value",
          "reasoning",
          "created_at",
          "supersedes_id",
          "revision_number",
          "origin"
        ],
        "title": "PersonalAssessmentResponse",
        "type": "object"
      },
      "PersonalAssessmentRevisionPayload": {
        "additionalProperties": false,
        "properties": {
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "value"
        ],
        "title": "PersonalAssessmentRevisionPayload",
        "type": "object"
      },
...
      "RestorePayload": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "target_status": {
            "anyOf": [
              {
                "enum": [
                  "new",
                  "to_review",
                  "interesting",
                  "shortlisted",
                  "deferred",
                  "excluded",
                  "archived"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Status"
          }
        },
        "title": "RestorePayload",
        "type": "object"
      },
...
      "StatusPayload": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "enum": [
              "new",
              "to_review",
              "interesting",
              "shortlisted",
              "deferred",
              "excluded",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusPayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Vocation",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/opportunities": {
      "get": {
        "operationId": "list_opportunities_api_opportunities_get",
        "parameters": [
          {
            "in": "query",
            "name": "group_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OpportunityListItemResponse"
                  },
                  "title": "Response List Opportunities Api Opportunities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Opportunities",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/opportunities/{opportunity_id}": {
      "get": {
        "operationId": "opportunity_detail_api_opportunities__opportunity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity Detail",
        "tags": [
          "opportunities"
        ]
      }
    },
...
    "/api/opportunities/{opportunity_id}/assessments/personal": {
      "get": {
        "operationId": "personal_assessments_api_opportunities__opportunity_id__assessments_personal_get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonalAssessmentResponse"
                  },
                  "title": "Response Personal Assessments Api Opportunities  Opportunity Id  Assessments Personal Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Personal Assessments",
        "tags": [
          "opportunities"
        ]
      },
      "post": {
        "operationId": "create_personal_assessment_api_opportunities__opportunity_id__assessments_personal_post",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalAssessmentPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalAssessmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Personal Assessment",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/opportunities/{opportunity_id}/assessments/personal/history": {
      "get": {
        "operationId": "personal_assessment_history_api_opportunities__opportunity_id__assessments_personal_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonalAssessmentResponse"
                  },
                  "title": "Response Personal Assessment History Api Opportunities  Opportunity Id  Assessments Personal History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Personal Assessment History",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/opportunities/{opportunity_id}/assessments/personal/{assessment_id}/revisions": {
      "post": {
        "operationId": "revise_personal_assessment_api_opportunities__opportunity_id__assessments_personal__assessment_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalAssessmentRevisionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalAssessmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revise Personal Assessment",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/opportunities/{opportunity_id}/decisions": {
      "get": {
        "operationId": "decision_history_api_opportunities__opportunity_id__decisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DecisionResponse"
                  },
                  "title": "Response Decision History Api Opportunities  Opportunity Id  Decisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decision History",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/opportunities/{opportunity_id}/exclude": {
      "post": {
        "operationId": "exclude_api_opportunities__opportunity_id__exclude_post",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExclusionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exclude",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/opportunities/{opportunity_id}/restore": {
      "post": {
        "operationId": "restore_api_opportunities__opportunity_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestorePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/opportunities/{opportunity_id}/status": {
      "post": {
        "operationId": "change_status_api_opportunities__opportunity_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Status",
        "tags": [
          "opportunities"
        ]
      }
    },
...
  }
}